# swagger.yaml
openapi: 3.0.0
info:
  title: My API
  version: 1.0.0
  description: API Documentation
servers:
  - url: http://localhost:8080
    description: HTTP Server
...
paths:
...
  /enrollment:
    get:
      tags:
        - Enrollments
      summary: Получить список заявок пользователя
      description: Возвращает список заявок на услуги для текущего пользователя.
      responses:
        '200':
          description: Успешный ответ
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Перенаправление на страницу входа
          headers:
            Location:
              schema:
                type: string
              example: "/auth/login"
        '403':
          description: Доступ запрещен (для мастеров и админов)
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Ошибка сервера
          content:
            text/plain:
              schema:
                type: string
...
  /services:
    get:
      tags:
        - Services
      summary: Получить все услуги
      description: Возвращает список всех услуг с информацией о мастерах и типах услуг.
      responses:
        '200':
          description: Успешный ответ
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Ошибка сервера
          content:
            text/plain:
              schema:
                type: string
...
  /profile/enrollment:
    get:
      tags:
        - Profile
      summary: Получить страницу заказов
      description: Возвращает страницу с заказами мастера.
      responses:
        '200':
          description: Успешный ответ
          content:
            text/html:
              schema:
                type: string
        '403':
          description: Необходима регистрация
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Мастер не найден
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Ошибка сервера
          content:
            text/plain:
              schema:
                type: string
...